Transmit the onClick function to the AddBook component
